Determine at compile time which memory pool to use
Memory pools are used for `sizeof (type)`-based allocations only.
The size of these allocations can be known at compile time, so it
is not necessary to pass that information to the pool allocator at
execution time. With the help of some macros, allocation can even
be directed to the appropriate (8 or 16 byte specific) memory pool
with no run time overhead.

JerryScript-DCO-1.0-Signed-off-by: Akos Kiss akiss@inf.u-szeged.hu